pyproject.toml
[tool.poetry]
name = "encode-templates"
version = "0.1.0"
description = ""
authors = ["dorgold46 <dgoldenberg@lightricks.com>"]
readme = "README.md"
packages = [{include = "encode_templates"}]

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^2.0.1"
google-cloud-bigquery = "^3.10.0"
pyyaml = "^6.0"
db-dtypes = "^1.1.1"
opencv-python = "^4.7.0.72"
torch = "^2.0.0"
torchvision = "^0.15.1"
torchaudio = "^2.0.1"
transformers = "^4.28.1"
scenedetect = "^0.6.1"
optimum = "^1.8.3"
scikit-learn = "^1.2.2"
matplotlib = "^3.7.1"
google-cloud-storage = "^2.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

